refactor: accordion codebase
