+++ package.json
@@ -47,7 +47,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.71.11",
     "react-native-gesture-handler": "~2.12.0",
-    "react-native-reanimated": "~2.17.0",
+    "react-native-reanimated": "~2.14.4",
     "react-native-svg": "13.9.0",
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.19.0",
